requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
spacy~=3.7.4
gensim~=4.3.2
pandas~=2.0.2
numpy~=1.23.5
bertopic~=0.16.1
nltk~=3.6.6
scikit-learn~=1.3.2
hdbscan~=0.8.33
spacytextblob~=4.0.0
typing~=3.6.4
pip~=23.3
pathlib~=1.0.1
emoji~=2.2.0
attrs~=23.1.0
wheel~=0.40.0
Pillow~=10.0.1
Jinja2~=3.1.2
setuptools~=65.5.1
packaging~=23.1
zipp~=3.15.0
Cython~=0.29.37
future~=0.18.3
platformdirs~=3.8.0
tqdm~=4.65.0
elasticsearch~=8.12.1
matplotlib~=3.7.5
networkx~=3.1
ipython~=8.12.3
pyvis~=0.3.2
seaborn~=0.13.2
scipy~=1.10.1
python-dotenv~=1.0.1
jupyter~=1.0.0
tornado~=6.4